+++ src/inspect_ai/util/_sandbox/environment.py
@@ -27,6 +27,10 @@ logger = logging.getLogger(__name__)
 ST = TypeVar("ST", bound="SandboxEnvironment")
 
 TaskInit = Callable[[str, Union["SandboxEnvironmentConfigType", None]], Awaitable[None]]
+TaskInitEnvironment = Callable[
+    [Union["SandboxEnvironmentConfigType", None], dict[str, str]],
+    Awaitable[dict[str, str]],
+]
 TaskCleanup = Callable[
     [str, Union["SandboxEnvironmentConfigType", None], bool], Awaitable[None]
 ]
@@ -237,6 +241,30 @@ class SandboxEnvironment(abc.ABC):
           config: Implementation defined configuration (optional).
         """
         pass
+
+    @classmethod
+    async def task_init_environment(
+        cls, config: SandboxEnvironmentConfigType | None, metadata: dict[str, str]
+    ) -> dict[str, str]:
+        """Called at task startup to identify environment variables required by task_init for a sample.
+
+        Return 1 or more environment variables to request a dedicated call to task_init
+        for samples that have exactly these environment variables (by default there is
+        only one call to task_init for all of the samples in a task if they share a
+        sandbox configuration).
+
+        This is useful for situations where config files are dynamic (e.g. through
+        sample metadata variable interpolation) and end up yielding different images
+        that need their own init (e.g. 'docker pull').
+
+        Args:
+            config: Implementation defined configuration (optional).
+            metadata: metadata: Sample `metadata` field
+
+        Returns:
+            Environment variables to set for call to task_init.
+        """
+        return {}
 
     @classmethod
     async def sample_init(